fix(discovery): multicast 组加入按平台拆分 — Windows 构建/测试恢复可用（#321）
multicast.go 的 reuseJoinControl 直接调用 syscall.SetsockoptInt/SetsockoptIPMreq
（int fd 签名，仅 Unix 成立）且无 build tag，Windows 上整个依赖 discovery 的
传递链（cmd/server、cmd/agent、internal/api/*）全部编译失败，本地全量
go test ./... 不可用。

修复：
- socket 组加入（SO_REUSEADDR + IP_ADD_MEMBERSHIP）拆入
  multicast_control_unix.go（//go:build !windows，原实现原样迁移）
  与 multicast_control_windows.go（//go:build windows，返回明确错误的 stub —
  ListenPacket 失败后 MulticastSource 按既有降级路径跳过 mDNS/SSDP 监听并告警，
  其余发现源与引擎不受影响；Windows 是开发平台，部署目标是 Linux）
- CI 新增 windows-build job：linux runner 交叉编译 GOOS=windows 的
  go vet + go build ./...（仅构建不跑测试），防止再次悄悄破坏

验证：linux go build/test ./... 全绿；GOOS=windows go vet+build ./... 通过；
GOOS=darwin go build ./internal/... ./cmd/... 通过（unix 实现双平台确认）

Closes #321
